Continuando implementação de upload de arquivo.

+++ src/main/java/br/ueg/loja/service/impl/ComputadorServiceImpl.java
@@ -3,6 +3,7 @@ package br.ueg.loja.service.impl;
 import br.ueg.loja.exception.SistemaMessageCode;
 import br.ueg.loja.model.Computador;
 import br.ueg.loja.repository.ComputadorRepository;
+import br.ueg.loja.repository.VendaRepository;
 import br.ueg.loja.service.ComputadorService;
 import br.ueg.prog.webi.api.exception.BusinessException;
 import org.apache.logging.log4j.util.Strings;
@@ -17,6 +18,8 @@ import java.util.Objects;
 public class ComputadorServiceImpl implements ComputadorService {
     @Autowired
     private ComputadorRepository computadorRepository;
+    @Autowired
+    private VendaRepository vendaRepository;
     @Override
     public Computador incluir(Computador computador) {
         this.validarCamposObrigatorios(computador);
@@ -79,7 +82,9 @@ public class ComputadorServiceImpl implements ComputadorService {
     @Override
     public Computador excluir(Long id) {
         Computador computadorExcluir = this.recuperarComputadorOuGeraErro(id);
-        this.computadorRepository.delete(computadorExcluir);
+        if (vendaRepository.count(computadorExcluir.getId()) == 0)
+            this.computadorRepository.delete(computadorExcluir);
+        else throw new BusinessException(SistemaMessageCode.ERRO_EXISTE_VENDA);
         return computadorExcluir;
     }
